# Tools and Tunings to make Kokkos GPU perf better