--- a/cpp/rrtmgp/kernels/mo_gas_optics_kernels.cpp
+++ b/cpp/rrtmgp/kernels/mo_gas_optics_kernels.cpp
@@ -18,7 +18,7 @@ void interpolation(int ncol, int nlay, int ngas, int nflav, int neta, int npres,
 
   // for (int ilay=1; ilay<=nlay; ilay++) {
   //   for (int icol=1; icol<=ncol; icol++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<2>(nlay,ncol) , YAKL_LAMBDA (int ilay, int icol) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<2>(nlay,ncol) , YAKL_LAMBDA (int ilay, int icol) {
     // index and factor for temperature interpolation
     jtemp(icol,ilay) = (int) ((tlay(icol,ilay) - (temp_ref_min - temp_ref_delta)) / temp_ref_delta);
     jtemp(icol,ilay) = std::min(ntemp - 1, std::max(1, jtemp(icol,ilay))); // limit the index range
@@ -31,13 +31,13 @@ void interpolation(int ncol, int nlay, int ngas, int nflav, int neta, int npres,
 
     // determine if in lower or upper part of atmosphere
     tropo(icol,ilay) = log(play(icol,ilay)) > press_ref_trop_log;
-  });
+  }));
 
   // for (int ilay=1; ilay<=nlay; ilay++) {
   //   for (int icol=1; icol<=ncol; icol++) {
   //     for (int iflav=1; iflav<=nflav; iflav++) {   // loop over implemented combinations of major species
   //       for (int itemp=1; itemp<=2; itemp++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<4>(nlay,ncol,nflav,2) , YAKL_LAMBDA (int ilay, int icol, int iflav , int itemp) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<4>(nlay,ncol,nflav,2) , YAKL_LAMBDA (int ilay, int icol, int iflav , int itemp) {
     yakl::FSArray<int,1,SB<2>> igases;
 
     // itropo = 1 lower atmosphere; itropo = 2 upper atmosphere
@@ -65,7 +65,7 @@ void interpolation(int ncol, int nlay, int ngas, int nflav, int neta, int npres,
     fmajor(2,1,itemp,iflav,icol,ilay) = (1. - fpress(icol,ilay)) * fminor(2,itemp,iflav,icol,ilay);
     fmajor(1,2,itemp,iflav,icol,ilay) =       fpress(icol,ilay)  * fminor(1,itemp,iflav,icol,ilay);
     fmajor(2,2,itemp,iflav,icol,ilay) =       fpress(icol,ilay)  * fminor(2,itemp,iflav,icol,ilay);
-  });
+  }));
 }
 
 
@@ -86,7 +86,7 @@ void combine_and_reorder_2str(int ncol, int nlay, int ngpt, real3d const &tau_ab
   //     for (int tgpt=1; tgpt<=gptTiles; tgpt++) {
   //       for (int itcol=1; itcol<=TILE_SIZE; itcol++) {
   //         for (int itgpt=1; itgpt<=TILE_SIZE; itgpt++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<5>(nlay,colTiles,gptTiles,TILE_SIZE,TILE_SIZE) , YAKL_LAMBDA (int ilay, int tcol, int tgpt, int itcol, int itgpt) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<5>(nlay,colTiles,gptTiles,TILE_SIZE,TILE_SIZE) , YAKL_LAMBDA (int ilay, int tcol, int tgpt, int itcol, int itgpt) {
     int icol = (tcol-1)*TILE_SIZE + itcol;
     int igpt = (tgpt-1)*TILE_SIZE + itgpt;
 
@@ -100,7 +100,7 @@ void combine_and_reorder_2str(int ncol, int nlay, int ngpt, real3d const &tau_ab
         ssa(icol,ilay,igpt) = 0.;
       }
     }
-  });
+  }));
 }
 
 
@@ -128,7 +128,7 @@ void compute_Planck_source(int ncol, int nlay, int nbnd, int ngpt, int nflav, in
   // for (int icol=1; icol<=ncol; icol++) {
   //   for (int ilay=1; ilay<=nlay; ilay++) {
   //     for (int igpt=1; igpt<=ngpt; igpt++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(nlay,ncol,ngpt) , YAKL_LAMBDA (int ilay, int icol, int igpt) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(nlay,ncol,ngpt) , YAKL_LAMBDA (int ilay, int icol, int igpt) {
     // itropo = 1 lower atmosphere; itropo = 2 upper atmosphere
     int itropo = merge(1,2,tropo(icol,ilay));  //WS moved itropo inside loop for GPU
     int iflav = gpoint_flavor(itropo, igpt); //eta interpolation depends on band's flavor
@@ -145,31 +145,31 @@ void compute_Planck_source(int ncol, int nlay, int nbnd, int ngpt, int nflav, in
                               fmajor(2,1,2,iflav,icol,ilay) * pfracin(igpt, jeta(2,iflav,icol,ilay)+1, jpress_loc-1, jtemp_loc+1) +
                               fmajor(1,2,2,iflav,icol,ilay) * pfracin(igpt, jeta(2,iflav,icol,ilay)  , jpress_loc  , jtemp_loc+1) +
                               fmajor(2,2,2,iflav,icol,ilay) * pfracin(igpt, jeta(2,iflav,icol,ilay)+1, jpress_loc  , jtemp_loc+1) );
-  });
+  }));
 
   //
   // Planck function by band for the surface
   // Compute surface source irradiance for g-point, equals band irradiance x fraction for g-point
   //
   // for (int icol=1; icol<=ncol; icol++) {
-  parallel_for( YAKL_AUTO_LABEL() , ncol , YAKL_LAMBDA (int icol) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , ncol , YAKL_LAMBDA (int icol) {
     interpolate1D(tsfc(icol), temp_ref_min, totplnk_delta, totplnk, planck_function.slice<1>(COLON,1,icol),nPlanckTemp,nbnd);
-  });
+  }));
   //
   // Map to g-points
   //
   // for (int igpt=1; igpt<=ngpt; igpt++) {
   //   for (int icol=1; icol<=ncol; icol++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<2>(ngpt,ncol) , YAKL_LAMBDA (int igpt, int icol) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<2>(ngpt,ncol) , YAKL_LAMBDA (int igpt, int icol) {
     sfc_src(igpt,icol) = pfrac(igpt,sfc_lay,icol) * planck_function(gpoint_bands(igpt), 1, icol);
-  });
+  }));
 
   // for (int icol=1; icol<=ncol; icol++) {
   //   for (int ilay=1; ilay<=nlay; ilay++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<2>(ncol,nlay) , YAKL_LAMBDA (int icol, int ilay) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<2>(ncol,nlay) , YAKL_LAMBDA (int icol, int ilay) {
     // Compute layer source irradiance for g-point, equals band irradiance x fraction for g-point
     interpolate1D(tlay(icol,ilay), temp_ref_min, totplnk_delta, totplnk, planck_function.slice<1>(COLON,ilay,icol),nPlanckTemp,nbnd);
-  });
+  }));
   //
   // Map to g-points
   //
@@ -179,21 +179,21 @@ void compute_Planck_source(int ncol, int nlay, int nbnd, int ngpt, int nflav, in
   // for (int icol=1; icol<=ncol; icol++) {
   //   for (int ilay=1; ilay<=nlay; ilay++) {
   //     for (int igpt=1; igpt<=ngpt; igpt++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(ncol,nlay,ngpt) , YAKL_LAMBDA (int icol, int ilay, int igpt) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(ncol,nlay,ngpt) , YAKL_LAMBDA (int icol, int ilay, int igpt) {
     lay_src(igpt,ilay,icol  ) = pfrac(igpt,ilay,icol  ) * planck_function(gpoint_bands(igpt),ilay,icol);
-  });
+  }));
 
   // compute level source irradiances for each g-point, one each for upward and downward paths
   // for (int icol=1; icol<=ncol; icol++) {
-  parallel_for( YAKL_AUTO_LABEL() , ncol , YAKL_LAMBDA (int icol) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , ncol , YAKL_LAMBDA (int icol) {
     interpolate1D(tlev(icol,1), temp_ref_min, totplnk_delta, totplnk, planck_function.slice<1>(COLON,1,icol),nPlanckTemp,nbnd);
-  });
+  }));
 
   // for (int icol=1; icol<=ncol; icol++) {
   //   for (int ilay=2; ilay<=nlay+1; ilay++) {
-  parallel_for( YAKL_AUTO_LABEL() , Bounds<2>(ncol,{2,nlay+1}) , YAKL_LAMBDA (int icol, int ilay) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , Bounds<2>(ncol,{2,nlay+1}) , YAKL_LAMBDA (int icol, int ilay) {
     interpolate1D(tlev(icol,ilay), temp_ref_min, totplnk_delta, totplnk, planck_function.slice<1>(COLON,ilay,icol),nPlanckTemp,nbnd);
-  });
+  }));
 
   //
   // Map to g-points
@@ -203,14 +203,14 @@ void compute_Planck_source(int ncol, int nlay, int nbnd, int ngpt, int nflav, in
   // for (int icol=1; icol<=ncol; icol+=2) {
   //   for (int ilay=1; ilay<=nlay; ilay++) {
   //     for (int igpt=1; igpt<=ngpt; igpt++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(ncol,nlay,ngpt) , YAKL_LAMBDA (int icol, int ilay, int igpt) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(ncol,nlay,ngpt) , YAKL_LAMBDA (int icol, int ilay, int igpt) {
     lev_src_dec(igpt,ilay,icol  ) = pfrac(igpt,ilay,icol  ) * planck_function(gpoint_bands(igpt),ilay,  icol  );
     lev_src_inc(igpt,ilay,icol  ) = pfrac(igpt,ilay,icol  ) * planck_function(gpoint_bands(igpt),ilay+1,icol  );
     if (icol < ncol) {
       lev_src_dec(igpt,ilay,icol+1) = pfrac(igpt,ilay,icol+1) * planck_function(gpoint_bands(igpt),ilay,  icol+1);
       lev_src_inc(igpt,ilay,icol+1) = pfrac(igpt,ilay,icol+1) * planck_function(gpoint_bands(igpt),ilay+1,icol+1);
     }
-  });
+  }));
 
   auto stop_t = std::chrono::high_resolution_clock::now();
   auto duration = std::chrono::duration_cast<std::chrono::microseconds>(stop_t - start_t);
@@ -233,7 +233,7 @@ void compute_tau_rayleigh(int ncol, int nlay, int nbnd, int ngpt, int ngas, int 
   // for (int ilay=1; ilay<=nlay; ilay++) {
   //   for (int icol=1; icol<=ncol; icol++) {
   //     for (int igpt=1; igpt<=ngpt; igpt++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(nlay,ncol,ngpt) , YAKL_LAMBDA (int ilay, int icol, int igpt) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(nlay,ncol,ngpt) , YAKL_LAMBDA (int ilay, int icol, int igpt) {
     int itropo = merge(1,2,tropo(icol,ilay)); // itropo = 1 lower atmosphere; itropo = 2 upper atmosphere
     int iflav = gpoint_flavor(itropo, igpt);
     // Inlining interpolate2D
@@ -242,7 +242,7 @@ void compute_tau_rayleigh(int ncol, int nlay, int nbnd, int ngpt, int ngas, int 
              fminor(1,2,iflav,icol,ilay) * krayl(igpt, jeta(2,iflav,icol,ilay)  , jtemp(icol,ilay)+1,itropo) +
              fminor(2,2,iflav,icol,ilay) * krayl(igpt, jeta(2,iflav,icol,ilay)+1, jtemp(icol,ilay)+1,itropo);
     tau_rayleigh(igpt,ilay,icol) =  k * (col_gas(icol,ilay,idx_h2o)+col_dry(icol,ilay));
-  });
+  }));
 }
 
 
@@ -343,7 +343,7 @@ void compute_tau_rayleigh(int ncol, int nlay, int nbnd, int ngpt, int ngas, int 
     // for (int ilay=1; ilay<=nlay; ilay++) {
     //   for (int icol=1; icol<=ncol; icol++) {
     //     for (int igpt0=0; igpt0<=max_gpt_diff; igpt0++) {
-    parallel_for( YAKL_AUTO_LABEL() , Bounds<2>(nlay,ncol) , YAKL_LAMBDA (int ilay, int icol) {
+    TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , Bounds<2>(nlay,ncol) , YAKL_LAMBDA (int ilay, int icol) {
       // This check skips individual columns with no pressures in range
       if ( layer_limits(icol,1) <= 0 || ilay < layer_limits(icol,1) || ilay > layer_limits(icol,2) ) {
       } else {
@@ -391,7 +391,7 @@ void compute_tau_rayleigh(int ncol, int nlay, int nbnd, int ngpt, int ngas, int 
           }
         }
       }
-    });
+    }));
   }
 
 #endif
@@ -411,7 +411,7 @@ void gas_optical_depths_major(int ncol, int nlay, int nbnd, int ngpt, int nflav,
   //   for (int icol=1; icol<=ncol; icol++) {
   //     // optical depth calculation for major species
   //     for (int igpt=1; igpt<=ngpt; igpt++) {
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(nlay,ncol,ngpt) , YAKL_LAMBDA (int ilay, int icol, int igpt) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(nlay,ncol,ngpt) , YAKL_LAMBDA (int ilay, int icol, int igpt) {
     // itropo = 1 lower atmosphere; itropo = 2 upper atmosphere
     int itropo = merge(1,2,tropo(icol,ilay));  // WS: moved inside innermost loop
 
@@ -432,7 +432,7 @@ void gas_optical_depths_major(int ncol, int nlay, int nbnd, int ngpt, int nflav,
                                                     fmajor(2,2,2,iflav,icol,ilay) * kmajor(igpt, jeta(2,iflav,icol,ilay)+1, jpress_loc  , jtemp_loc+1) );
 
     tau(igpt,ilay,icol) += tau_major;
-   });
+   }));
 }
 
 
@@ -461,7 +461,7 @@ void compute_tau_absorption(int max_gpt_diff_lower, int max_gpt_diff_upper, int 
   if (top_at_1) {
 
     // for (int icol=1; icol<=ncol; icol++){
-    parallel_for( YAKL_AUTO_LABEL() , ncol , YAKL_LAMBDA (int icol) {
+    TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , ncol , YAKL_LAMBDA (int icol) {
       itropo_lower(icol,2) = nlay;
       // itropo_lower(icol,1) = minloc(play(icol,:), dim=1, mask=tropo(icol,:))
       {
@@ -493,12 +493,12 @@ void compute_tau_absorption(int max_gpt_diff_lower, int max_gpt_diff_upper, int 
         }
         itropo_upper(icol,2) = maxloc;
       }
-    });
+    }));
 
   } else {  // top_at_1
 
     // for (int icol=1; icol<=ncol; icol++){
-    parallel_for( YAKL_AUTO_LABEL() , ncol , YAKL_LAMBDA ( int icol ) {
+    TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , ncol , YAKL_LAMBDA ( int icol ) {
       itropo_lower(icol,1) = 1;
       // itropo_lower(icol,2) = minloc(play(icol,:), dim=1, mask=tropo(icol,:))
       {
@@ -531,7 +531,7 @@ void compute_tau_absorption(int max_gpt_diff_lower, int max_gpt_diff_upper, int 
         itropo_upper(icol,1) = maxloc;
       }
 
-    });
+    }));
 
   }  // top_at_1
 
@@ -573,7 +573,7 @@ void combine_and_reorder_nstr(int ncol, int nlay, int ngpt, int nmom, real3d con
   // do icol = 1, ncol
   //   do ilay = 1, nlay
   //     do igpt = 1, ngpt
-  parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(ncol,nlay,ngpt) , YAKL_LAMBDA (int icol, int ilay, int igpt) {
+  TIMED_KERNEL(parallel_for( YAKL_AUTO_LABEL() , SimpleBounds<3>(ncol,nlay,ngpt) , YAKL_LAMBDA (int icol, int ilay, int igpt) {
     real t = tau_abs(igpt,ilay,icol) + tau_rayleigh(igpt,ilay,icol);
     tau(icol,ilay,igpt) = t;
     if (t > 2. * tiny) {
@@ -587,6 +587,6 @@ void combine_and_reorder_nstr(int ncol, int nlay, int ngpt, int nmom, real3d con
     if (nmom >= 2) {
       p(icol,ilay,igpt,2) = 0.1;
     }
-  });
+  }));
 }
 #endif