--- a/cpp/rrtmgp/kernels/mo_gas_optics_kernels.cpp
+++ b/cpp/rrtmgp/kernels/mo_gas_optics_kernels.cpp
@@ -117,8 +117,6 @@ void compute_Planck_source(int ncol, int nlay, int nbnd, int ngpt, int nflav, in
   using yakl::fortran::SimpleBounds;
   using yakl::fortran::Bounds;
 
-  auto start_t = std::chrono::high_resolution_clock::now();
-
   real3d pfrac          ("pfrac"          ,ngpt,nlay,ncol);
   real3d planck_function("planck_function",nbnd,nlay+1,ncol);
   real1d one            ("one"            ,2);
@@ -211,15 +209,7 @@ void compute_Planck_source(int ncol, int nlay, int nbnd, int ngpt, int nflav, in
       lev_src_inc(igpt,ilay,icol+1) = pfrac(igpt,ilay,icol+1) * planck_function(gpoint_bands(igpt),ilay+1,icol+1);
     }
   }));
-
-  auto stop_t = std::chrono::high_resolution_clock::now();
-  auto duration = std::chrono::duration_cast<std::chrono::microseconds>(stop_t - start_t);
-  static double total_time = 0.0;
-  total_time += (duration.count() / 1000000.0);
-  std::cout << "Total time spent: " << total_time << std::endl;
 }
-
-
 
 // compute Rayleigh scattering optical depths
 void compute_tau_rayleigh(int ncol, int nlay, int nbnd, int ngpt, int ngas, int nflav, int neta, int npres, int ntemp,
@@ -244,8 +234,6 @@ void compute_tau_rayleigh(int ncol, int nlay, int nbnd, int ngpt, int ngas, int 
     tau_rayleigh(igpt,ilay,icol) =  k * (col_gas(icol,ilay,idx_h2o)+col_dry(icol,ilay));
   }));
 }
-
-
 
 // compute minor species optical depths
 #ifdef RRTMGP_CPU_KERNELS