On weaver after opts with kokkos:
```
Kokkos:: Longwave  did 100 loops of 1000 cols and 42 layers in 0.989172 s
YAKL::   Longwave  did 100 loops of 1000 cols and 42 layers in 0.992474 s
Kokkos:: Shortwave did 100 loops of 1000 cols and 42 layers in 1.058807 s
YAKL::   Shortwave did 100 loops of 1000 cols and 42 layers in 1.085853 s
```

So, Kokkos is now slightly faster than YAKL on GPU. It was over 2x slower before these changes.

Methodology: A new `TIMED_KERNEL` macro enables timings for all kernels, both YAKL and Kokkos. When this is enabled, raw data is dumped. The new tool `rrtmgp-perf-analysis` can read these raw dumps and present the user with concise, useful information, such as which kernels were taking the most time and how their perf compared to their YAKL counterparts. For most kernels that were slower than YAKL, improving perf was just a matter of rearranging the MDrange order.